refactor: modified UserQuiz score data type

+++ module-api/src/main/java/com/codingbottle/domain/quiz/model/UserQuizInfo.java
@@ -4,7 +4,7 @@ public record UserQuizInfo(
         String nickname,
         int score
 ) {
-    public static UserQuizInfo from(String nickname, Double score) {
-        return new UserQuizInfo(nickname, score.intValue());
+    public static UserQuizInfo from(String nickname, int score) {
+        return new UserQuizInfo(nickname, score);
     }
 }
